proj: implement vg_draw_char and fix a warning in keyboard.c
git-svn-id: https://svn.fe.up.pt/repos/lcom1213-t1g5@130 dd1ae04f-8964-4e22-877d-0b2b87089ac8

+++ proj/src/keyboard.c
@@ -55,7 +55,7 @@ char key_to_char(unsigned int scancode) {
     if (altgr)
         res = key_c[scancode][KEY_IDX_ALTGR];
     else if (up || shift) {
-        if (isalpha(key_c[scancode][KEY_IDX_NONE])) {
+        if (isalpha((int)(key_c[scancode][KEY_IDX_NONE]))) {
             if (up)
                 res = key_c[scancode][KEY_IDX_UP];
             else if (shift)
